Added submission page
I've been working to totally reimplement the pagination system.  Having to manually
name files like 1.txt, 1.png, etc. is extremely user-unfriendly.  So I've
added a database to store all the files to read in a database which is
then read from.  I've also begun work on an adding page, which currently
only previews the submission and has no mechanism to submit the page.